Update IMT.csv
Add imt code for cyclone wind speed 10-min sustained. It was already there as km/h but STORM uses m/s...

This makes me reconsider the discussion in #5, and splitting IMT into two separate codelists:
- metric (optional)
- unit (mandatory)

So even if the specific metric (larger range of variability) is not in the codelist, one can still indicate the unit if part of the most common ones (smaller range of variability).